replace int with size_t.

+++ src/relay/op/tensor/transform.cc
@@ -254,7 +254,7 @@ bool ConcatenateRel(const Array<Type>& types,
     if (e_dtype != dtype) {
       throw relay::Error("relay.concatenate requires all tensors have the same dtype");
     }
-    for (int j = 0; j < first->shape.size(); ++j) {
+    for (size_t j = 0; j < first->shape.size(); ++j) {
       if (j == param->axis) continue;
       if (reporter->AssertEQ(first->shape[j], e->shape[j])) continue;
       throw relay::Error("relay.concatenate requires all tensors have the same shape "
@@ -395,8 +395,8 @@ bool StackRel(const Array<Type>& types,
     const DataType& e_dtype = e->dtype;
     CHECK_EQ(e_ndim, ndim) << "relay.stack requires all tensors have the same ndim";
     CHECK_EQ(e_dtype, dtype) << "relay.stack requires all tensors have the same dtype";
-    for (int j = 0; j < static_cast<int>(first->shape.size()); ++j) {
-      if (j == int(param->axis)) continue;
+    for (size_t j = 0; j < first->shape.size(); ++j) {
+      if (j == param->axis) continue;
       if (reporter->AssertEQ(first->shape[j], e->shape[j])) continue;
       throw relay::Error("relay.stack requires all tensors have the same shape "
                          "on non-stacking axes");
